grouped the data

+++ app/src/main/java/com/vicky/apps/datapoints/ui/model/DataFields.kt
@@ -3,6 +3,6 @@ package com.vicky.apps.datapoints.ui.model
 data class DataFields(
 
     val year: String,
-    val volume_of_mobile_data: Int,
+    val volume_of_mobile_data: Double,
     val lowest_quarter:String
 )
